Fix misguided type change from ac4d3aff
This is a String, not a std::string. Changing to size_t may have silenced the
compiler warning, but that doesn't mean it's correct. Neither was the original
`int`. Of course Arduino doesn't document the return type of .length at
https://www.arduino.cc/reference/en/language/variables/data-types/string/functions/length
but the definition in WString.h is clear.

+++ WiFiSettings.cpp
@@ -47,7 +47,7 @@ String pwgen() {
 }
 String html_entities(const String& raw) {
     String r;
-    for (size_t i = 0; i < raw.length(); i++) {
+    for (unsigned int i = 0; i < raw.length(); i++) {
         char c = raw.charAt(i);
         if (c >= '!' && c <= 'z' && c != '&' && c != '<' && c != '>') {
             // printable ascii minus html and {}
